requirements.txt
Django==5.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-filter==23.5
djangorestframework==3.14.0
Pillow==10.2.0
python-dotenv==1.0.1
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.1
python-dateutil==2.8.2
requests==2.31.0 